+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Version X
+  - support SAN for client certificates
+
 # Version 1.2.1
   - run slashing protection database garbage collection periodically
   - add commit hash to log on startup